explorer: refactor cancel question feedback
